Harshad | Some restructuring
