+++ README.md
@@ -1,22 +1,14 @@
 # polaris
 
+An extremely light weight workflow orchestrator for Golang.
+
 Inspired from https://github.com/flipkart-incubator/databuilderframework
-
-## Introduction
-
-The polaris framework is a high level logic execution engine in Golang that can be used to execute multi-step/long running workflows. 
-
-#### When should you use Polaris?
-
+   
+## Use cases
 1. You have multi-step workflow executions where each step is dependent on data generated from previous steps.
 2. Executions can span one request scope or multiple scopes.
 3. Your systems works with reusable components that can be combined in different ways to generate different end-results.
 4. Your workflows can pause, resume or even restart from the beginning.
-   
-## Use cases
-1. An OMS for any sort of checkout journey (Zerodha, Uber, Swiggy).
-2. Ticket processing flows.
-3. Image / Media processing, etc.
 
 ## Limitations
 1. Workflow versioning is tricky to implement:
@@ -43,10 +35,6 @@ The polaris framework is a high level logic execution engine in Golang that can 
     * **Transients** - A set of names of Data that would be considered Transients for this case. (See later for a detailed explanation of transients)
 * _**ExecutionGraph**_ - A graph of connected and topologically sorted builders that are used by the execution engine to execute a flow. 
 * _**DataSet**_ - A set of the Data provided by the client and generated internally by the different builders of a particular ExecutionGraph
-* _**DataFlowInstance**_ - An instantiation of DataFlow that contains it's own copy of ExecutionGraph and DataSet. This represents the execution context of a particular request.
-* _**DataDelta**_ - The set of new data that needs to be considered as input for a particular execution
-* _**DataFlowExecutor**_ - The core engine that uses the provided DataDelta to execute the ExecutionGraph present in the given DataFlowInstance. This will augment the DataSet within the DataFlowInstance with the non-transient Data generated by the different Builders. All newly generated Data by the engine (including generated transient Data) is returned.
-* _**DataSetAccessor**_ - A typesafe utility for accesing Data present in a DataSet. This is used inside the Builders to generate Data.
 
 ## How does the framework perform at scale?
 The framework itself has extremely low overhead. Since execution graphs are generated pre-runtime, all the orchestrator will do at runtime is use the graph and available data to run whichever builders can be run. Benchmarking pending.
