+++ preprocessor.go
@@ -2,6 +2,9 @@ package polaris
 
 import (
 	"container/list"
+	"reflect"
+
+	mapset "github.com/deckarep/golang-set/v2"
 )
 
 var workflowStore map[string]DataFlow
@@ -11,6 +14,50 @@ func init() {
 	workflowStore = make(map[string]DataFlow)
 	dataStore = make(map[string]DataSet)
 	// This is where redis client would be initialized
+}
+
+type MetaDataManager struct {
+	builders              map[string]IBuilder
+	builderMetaMap        map[string]BuilderMeta
+	producedToProducerMap map[string]BuilderMeta // TODO: make this []BuilderMeta when ResolutionSpec is implemented
+}
+
+func newMetaDataManager() MetaDataManager {
+	return MetaDataManager{
+		builders:              make(map[string]IBuilder),
+		builderMetaMap:        make(map[string]BuilderMeta),
+		producedToProducerMap: make(map[string]BuilderMeta),
+	}
+}
+
+func buildSet(data []IData) mapset.Set[string] {
+	set := mapset.NewSet[string]()
+	for _, d := range data {
+		set.Add(Name(d))
+	}
+	return set
+}
+
+func newBuilderMeta(builder IBuilder) BuilderMeta {
+	builderInfo := builder.GetBuilderInfo()
+	return BuilderMeta{
+		Consumes:  buildSet(builderInfo.Consumes),
+		Optionals: buildSet(builderInfo.Optionals),
+		Accesses:  buildSet(builderInfo.Accesses),
+		Produces:  Name(builderInfo.Produces),
+		Name:      Name(builder),
+		Type:      reflect.TypeOf(builder),
+	}
+}
+
+func (m *MetaDataManager) register(builder IBuilder) {
+	builderMeta := newBuilderMeta(builder)
+	if _, ok := m.builderMetaMap[builderMeta.Name]; ok {
+		panic("Builder already exists")
+	}
+	m.builders[builderMeta.Name] = builder
+	m.builderMetaMap[builderMeta.Name] = builderMeta
+	m.producedToProducerMap[builderMeta.Produces] = builderMeta
 }
 
 type DataFlow struct {
