--- workflow.go
@@ -1,10 +1,0 @@
-package polaris
-
-type IWorkflow interface {
-	GetWorkflowMeta() WorkflowMeta
-}
-
-type WorkflowMeta struct {
-	Builders   []IBuilder
-	TargetData IData
-}
